+++ glx/lib/gedcom_converter.go
@@ -150,7 +150,7 @@ func (conv *ConversionContext) Convert(records []*GEDCOMRecord) error {
 
 				relationship := &Relationship{
 					Type: relType,
-					Participants: []RelationshipParticipant{
+					Participants: []Participant{
 						{Person: parentID, Role: ParticipantRoleParent},
 						{Person: link.PersonID, Role: ParticipantRoleChild},
 					},